ThymusABM.jl Documentation

ThymusABM.jl Documentation

Functions

Index

Instructions to run ThymusABM.jl

Introduction

Functions

ThymusABM.Dendritic — Type

Dendritic
id::Int                             # Unique ID to identify agent
pos::NTuple{3,Float64}              # 2D position of agent
vel::NTuple{3,Float64}              # 2D velocity of agent
mass::Float64                       # Mass of agent to use in collisions
type::Symbol                        # Cell type of agent
color::String                       # Color used for agent in videos
size::Float64                       # Size of agent in videos
num_interactions::Int               # Total number of interactions agent has
antigens::Array                     # List of antigens the Dendritic agent contains
steps_alive::Int                    # Number of steps the Dendritic agent has been alive for

Dendritic agent struct.

ThymusABM.Parameters — Type

Parameters
width_height::NTuple{3,Float64} = width_height                          # 3-Tuple of Float64 for 3 dimensions of the simulation (x, y, z)
speed::Float64 = speed                                                  # Float64 speed factor
dt::Float64 = dt                                                        # Float64 of time step governing movement/velocity of agents
n_tecs::Int = n_tecs                                                    # Int of number of mTECs to have in simulation
n_thymocytes::Int = n_thymocytes                                        # Int of number of thymocytes to have in simulation
n_dendritics::Int = n_dendritics                                        # Int of number of DCs to have in simulation
threshold::Int16 = threshold                                            # Int of negative selection threshold of thymocytes
peptides::Array = peptides                                              # Array of all possible peptides a mTEC/DC can present
tcrs::Array = tcrs                                                      # Array of generated TCRs for thymocyte to randomly select from
successful_interactions::Int = successful_interactions                  # Int of number of times a thymocyte was selected in an interaction with mTEC/DC
unsuccessful_interactions::Int = unsuccessful_interactions              # Int of number of times a thymocyte was not selected in an interaction with mTEC/DC
escaped_thymocytes::Int = 0                                             # Int of number of thymocytes that were not negatively selected that should have been
autoreactive_thymocytes::Int = 0                                        # Int of number of thymocytes that were negatively selected or escaped
nonautoreactive_thymocytes::Int = 0                                     # Int of number of thymocytes that were not autoreactive and can be assumed to have successfully exited the thymus
total_thymocytes::Int = 0                                               # Int of total number of thymocytes that have entered simulation
treg_threshold::Float64 = treg_threshold                                # Float64 setting the threshold for a nonautoreactive thymocyte to be classified as a Treg
num_tregs::Int = 0                                                      # Int of total number of Tregs that have successfully exited simulation
tecs_present::Int = n_tecs                                              # Int of number of mTECs currently present in the simulation
aa_matrix::NamedArray = aa_matrix                                       # NamedArray containing all possible interaction strengths between peptides and tcrs
synapse_interactions::Int = synapse_interactions                        # Int of total number of peptide:TCR reactions to calculate for one thymocyte:mTEC/DC interaction
deaths::Int = deaths                                                    # Int of total number of thymocytes that have exited simulation for any reason at current step
total_dead_thymocytes::Int = total_dead_thymocytes                      # Int of total number of thymocytes that have exited simulation for any reason across all steps
alive_thymocytes::Int = alive_thymocytes                                # Int of total number of alive thymocytes at current step
stage_genes_peptides_dict::Vector{Any} = stage_genes_peptides_dict      # Vector holding the development stage information for mTECs (genes and corresponding peptides for each stage)
max_tec_interactions::Int = max_tec_interactions                        # Int of maximum number of interactions a mTEC can have before it "dies" and is replaced by a new one
max_thymocyte_interactions::Int = max_thymocyte_interactions            # Int of maximum number of interactions a thymocyte must have before it leaves as a nonautoreactive T cell
step::Int = 0                                                           # Int of current step of simulation
mtec_genes::Int = mtec_genes                                            # Int of number of genes mtec will express
mtec_peptides::Int = mtec_peptides                                      # Int of number of peptides mtec will present from its genes

Parameters struct.

ThymusABM.Tec — Type

Tec
id::Int                             # Unique ID to identify agent
pos::NTuple{3,Float64}              # 2D position of agent
vel::NTuple{3,Float64}              # 2D velocity of agent
mass::Float64                       # Mass of agent to use in collisions
type::Symbol                        # Cell type of agent
color::String                       # Color used for agent in videos
size::Float64                       # Size of agent in videos
num_interactions::Int               # Total number of interactions agent has
antigens::Array                     # List of antigens the Tec agent contains
stage::Int                          # Maturation stage of Tec
steps_alive::Int                    # Number of steps the Tec agent has been alive for

Tec agent struct.

ThymusABM.Thymocyte — Type

Thymocyte
id::Int                             # Unique ID to identify agent
pos::NTuple{3,Float64}              # 2D position of agent
vel::NTuple{3,Float64}              # 2D velocity of agent
mass::Float64                       # Mass of agent to use in collisions
type::Symbol                        # Cell type of agent
color::String                       # Color used for agent in videos
size::Float64                       # Size of agent in videos
num_interactions::Int               # Total number of interactions agent has
tcr::String                         # TCR that the thymocyte agent is carrying
reaction_levels::Dict               # Dict to hold thymocyte's seen antigens and reaction levels to them
death_label::Bool                   # Label a thymocyte to be killed on its next step
confined::Bool                      # Boolean if the thymocyte has been confined by an APC
bind_location::NTuple{3,Float64}    # Location of binding w/ APC that led to confinement
treg::Bool                          # Boolean if thymocyte meets Treg threshold or not
steps_alive::Int                    # Number of steps the Thymocyte agent has been alive for

Thymocyte agent struct.

ThymusABM.add_dendritics! — Method

add_dendritics!(model, n_dendritics, color, size)

Adds n_dendritics number of Dendritics to the model of given color and size.

ThymusABM.add_tecs! — Method

add_tecs!(model, n_tecs, color, size, replenishing)

Adds n_tecs number of Tecs to the model of given color and size. Boolean replenishing determines if Tec is an initial Tec (true), or added to the model later on (false).

ThymusABM.add_thymocytes! — Method

add_thymocytes!(model, n_thymocytes, color, size, initial)

Adds n_thymocytes number of Thymocytes to the model of given color and size. Boolean initial determines if Thymocyte is an initial Thymocyte (true), or added to the model later on (false).

ThymusABM.agent_peptides_encountered — Method

agent_peptides_encountered(agent)

Return the number of peptides that agent has encountered.

ThymusABM.alive_ratio — Method

alive_ratio(model)

Return the proportion of alive thymocytes to exited and dead thymocytes in the model.

ThymusABM.cell_colors — Method

cell_colors(a)

Return the color of the given agent a.

ThymusABM.cell_markers — Method

cell_markers(a)

Return the marker of the given agent a.

ThymusABM.cell_move! — Method

cell_move!(agent::Union{Tec, Dendritic, Thymocyte}, model)

Move the given agent of the model according to its type.

ThymusABM.cell_sizes — Method

cell_sizes(a)

Return the size of the given agent a.

ThymusABM.collide! — Method

collide!(a::AbstractAgent, b::AbstractAgent)

Calculates an elastic collision between two agents, a and b, when they collide with each other. Adapted from Agents.jl elastic_collision and modified to work for 3D animations.

ThymusABM.escape_ratio — Method

escape_ratio(model)

Return the proportion of exited and dead thymocytes that escaped in the model.

ThymusABM.escapedautoreactive_ratio — Method

escapedautoreactive_ratio(model)

Return the proportion of autoreactive thymocytes that escaped in the model.

ThymusABM.initialize — Method

initialize(;
    width_height = (1.0, 1.0, 1.0),
    speed = 0.002,
    n_tecs = 50,
    n_thymocytes = 1000,
    n_dendritics = 50,
    dt=1.0,
    threshold = 0.8,
    treg_threshold = 0.6,
    rng_seed = 1,
    synapse_interactions = 1,
    max_tec_interactions = 200,
    max_thymocyte_interactions = 80)

Initialize the model with default or specified parameters. Returns the initialized model.

ThymusABM.model_step! — Method

model_step!(model)

Steps the model one step forward. Calculates all agent interactions for the given step. Determines when an agent leaves due to age and adds new agent to replenish it.

ThymusABM.model_volume — Method

model_volume(model)

Return the volume of the model.

ThymusABM.nonreact_ratio — Method

nonreact_ratio(model)

Return the proportion of exited and dead thymocytes that were non-autoreactive in the model.

ThymusABM.react_ratio — Method

react_ratio(model)

Return the proportion of exited and dead thymocytes that were autoreactive in the model.

ThymusABM.read_parameters — Method

read_parameters(filename::String)

Obtain and return the model parameters from the config file named filename.

ThymusABM.run — Method

run()

Run the model.

ThymusABM.set_color! — Method

set_color!(agent::Union{Tec, Dendritic, Thymocyte}, model)

Set the color of the given agent of the model according to its type.

ThymusABM.tec — Method

tec(a)

Used to check if agent a is a Tec.

ThymusABM.tec_DC_interact! — Method

tec_DC_interact!(a1::Union{Tec, Dendritic, Thymocyte}, a2::Union{Tec, Dendritic, Thymocyte})

Governs the interaction between a Tec and Dendritic who come into contact. Transfers Tec antigens to Dendritic antigens.

ThymusABM.thymocyte — Method

thymocyte(a)

Used to check if agent a is a Thymocyte.

ThymusABM.thymocyte_APC_interact! — Method

thymocyte_APC_interact!(a1::Union{Tec, Dendritic, Thymocyte}, a2::Union{Tec, Dendritic, Thymocyte}, model)

Governs the interaction between a Thymocyte and Tec (or Dendritic) of the model who come into contact. Calculates Thymocyte TCR binding affinity to one of Tec or Dendritic's peptides.

ThymusABM.total_thy — Method

total_thy(model)

Return the total number of dead, exited, and alive thymocytes by end of simulation of the model.

ThymusABM.update_tec_stage — Method

update_tec_stage(tec, model)

Updates the development stage of the given Tec agent tec of the model according to how many interactions with Thymocytes it has had so far.

Index

ThymusABM.Dendritic

ThymusABM.Parameters

ThymusABM.Tec

ThymusABM.Thymocyte

ThymusABM.add_dendritics!

ThymusABM.add_tecs!

ThymusABM.add_thymocytes!

ThymusABM.agent_peptides_encountered

ThymusABM.alive_ratio

ThymusABM.cell_colors

ThymusABM.cell_markers

ThymusABM.cell_move!

ThymusABM.cell_sizes

ThymusABM.collide!

ThymusABM.escape_ratio

ThymusABM.escapedautoreactive_ratio

ThymusABM.initialize

ThymusABM.model_step!

ThymusABM.model_volume

ThymusABM.nonreact_ratio

ThymusABM.react_ratio

ThymusABM.read_parameters

ThymusABM.run

ThymusABM.set_color!

ThymusABM.tec

ThymusABM.tec_DC_interact!

ThymusABM.thymocyte

ThymusABM.thymocyte_APC_interact!

ThymusABM.total_thy

ThymusABM.update_tec_stage